Add copy-up-exclude option for rootlesskit
* add --copy-up-exclude flag for rootlesskit
* modify tmpfs+symlink driver to handle content exclusion

+++ pkg/copyup/tmpfssymlink/tmpfssymlink.go
@@ -15,7 +15,24 @@ func NewChildDriver() copyup.ChildDriver {
 	return &childDriver{}
 }
 
+func NewChildDriverWithExclusion(excludedFiles []string) copyup.ChildDriver {
+	return &childDriver{
+		excludedFiles,
+	}
+}
+
 type childDriver struct {
+	excludedFiles []string
+}
+
+func isExcludedFile(path string, dirs []string) bool {
+	for _, fExcluded := range dirs {
+		if path == fExcluded {
+			_, _ = fmt.Fprintf(os.Stderr, "matched: %s", path)
+			return true
+		}
+	}
+	return false
 }
 
 func (d *childDriver) CopyUp(dirs []string) ([]string, error) {
@@ -27,24 +44,24 @@ func (d *childDriver) CopyUp(dirs []string) ([]string, error) {
 	}
 	defer os.RemoveAll(bind0)
 	var copied []string
-	for _, d := range dirs {
-		d := filepath.Clean(d)
-		if d == "/tmp" {
+	for _, dir := range dirs {
+		dir := filepath.Clean(dir)
+		if dir == "/tmp" {
 			// TODO: we can support copy-up /tmp by changing bind0TempDir
 			return copied, errors.New("/tmp cannot be copied up")
 		}
 
-		if err := unix.Mount(d, bind0, "", uintptr(unix.MS_BIND|unix.MS_REC), ""); err != nil {
-			return copied, fmt.Errorf("failed to create bind mount on %s: %w", d, err)
+		if err := unix.Mount(dir, bind0, "", uintptr(unix.MS_BIND|unix.MS_REC), ""); err != nil {
+			return copied, fmt.Errorf("failed to create bind mount on %s: %w", dir, err)
 		}
 
-		if err := unix.Mount("none", d, "tmpfs", 0, ""); err != nil {
-			return copied, fmt.Errorf("failed to mount tmpfs on %s: %w", d, err)
+		if err := unix.Mount("none", dir, "tmpfs", 0, ""); err != nil {
+			return copied, fmt.Errorf("failed to mount tmpfs on %s: %w", dir, err)
 		}
 
-		bind1, err := os.MkdirTemp(d, ".ro")
+		bind1, err := os.MkdirTemp(dir, ".ro")
 		if err != nil {
-			return copied, fmt.Errorf("creating a directory under %s: %w", d, err)
+			return copied, fmt.Errorf("creating a directory under %s: %w", dir, err)
 		}
 		if err := unix.Mount(bind0, bind1, "", uintptr(unix.MS_MOVE), ""); err != nil {
 			return copied, fmt.Errorf("failed to move mount point from %s to %s: %w", bind0, bind1, err)
@@ -65,7 +82,10 @@ func (d *childDriver) CopyUp(dirs []string) ([]string, error) {
 			} else {
 				symlinkSrc = filepath.Join(filepath.Base(bind1), f.Name())
 			}
-			symlinkDst := filepath.Join(d, f.Name())
+			symlinkDst := filepath.Join(dir, f.Name())
+			if len(d.excludedFiles) > 0 && isExcludedFile(symlinkDst, d.excludedFiles) {
+				continue
+			}
 			// `mount` may create extra `/etc/mtab` after mounting empty tmpfs on /etc
 			// https://github.com/rootless-containers/rootlesskit/issues/45
 			if err = os.RemoveAll(symlinkDst); err != nil {
@@ -75,7 +95,7 @@ func (d *childDriver) CopyUp(dirs []string) ([]string, error) {
 				return copied, fmt.Errorf("symlinking %s to %s: %w", symlinkSrc, symlinkDst, err)
 			}
 		}
-		copied = append(copied, d)
+		copied = append(copied, dir)
 	}
 	return copied, nil
 }
